clean up type names
Summary:
This diff started because I was irritated in debugging that a DYNAMIC inside a Union
showed up as "builtins.object" instead of "dynamic", but it kinda grew into a full
cleanup of how we handle type names :) Changes made:

1. Eliminate the module name prefix from a type name when it is `builtins`, `__static__`,
or `typing`. This just makes things a lot less verbose and I think all of the names in
these modules are sufficiently unique and ubiquitous that it doesn't introduce any
confusion or ambiguity.

2. `GenericTypeName` defers to the `name` property of each type parameter when
constructing the full generic type name, instead of using the `type_name` of the type
directly. This change is what actually fixes my original irritation; the `TypeName` of
`DYNAMIC_TYPE` is `builtins.object` but it overrides its `name` property to `"dynamic"`;
now we respect that override even when it is inside a union.

3. We used to name types as e.g. `foo.C` and instances as `foo.C instance`. Since we
use instances most places, this led to a ton of `x.klass.name` instead of `x.name`
to avoid the kinda verbose and not-very-informative ` instance` suffix everywhere. In
this diff I switch to using the simpler `foo.C` for an instance of `foo.C`, and `Type[foo.C]`
for the actual type. This is more consistent with how those types look in annotations.
This leads to some cases of `x.instance.name` to avoid the `Type[]` wrapper in some
messages, but fewer than we used to have `x.klass.name`. And some of these should
go away as we move away from passing around `Class` to represent instances of a type.

4. The more accurate naming of generic type parameters uncovered a bug where in
a type annotation like `int | str` we resolved it to `Union[Exact[int], Exact[str]]` instead of
`Union[int, str]`. We call `inexact()` on annotations to avoid this, but `inexact()` and friends
didn't have support for recursively inexactifying unions. Now they do.

Reviewed By: DinoV

Differential Revision: D27679443

fbshipit-source-id: 903fa73cb8