support JUMP_IF_[NON]ZERO_OR_POP in the JIT
Summary:
At one point we thought we might get rid of these opcodes and
instead the JIT support for `JUMP_IF_{TRUE,FALSE}_OR_POP` could just
check its input type and do the right thing if the input is a primitive.
But I'm not sure we should do that. For one thing, our general approach
with primitives in every other area has been distinct opcodes; doing a
polymorphic opcode here seems inconsistent without clear rationale. And
secondly, "do the right thing" with the current HIR opcode set involves
emitting different HIR (IsTruthy plus CondBranch vs just CondBranch),
and we don't know the input type yet in the builder. So we'd either end
up rewriting HIR later, or adding a polymorphic version of CondBranch
that does the IfTruthy internally if needed. Neither of those seems
preferable to just continuing our pattern of separate less-polymorphic
opcodes.

This support was a bit more complex than expected because the typical
cases for `JUMP_IF_[NON]ZERO_OR_POP` are cases where we are using the
output of a compare as a value and as input to a branch. With
Python objects this doesn't occur in LIR, because the value on the stack
is a Python object and the output of `IsTruthy` on that object goes to the
condbranch; nothing else will ever use the output of `IsTruthy`. But
with static Python primitives the value on the stack is also directly
the input to the condbranch, and might also be used later. So we have to
account for this possibility in LIR.

The other change here is in the tests. The previous
`test_int_compare_or_*` tests here were not really safe and should
result in type errors in the static compiler, because they create
`Union[primitive_type, object_type]` types that can't be safely handled.
This actually broke once these tests were JITted because the JIT (rightly)
objected to returning a `TCInt32 | Object` from the function.
So I modified the test to do a type-safe `or` chain (and `and` chain)
instead. The follow-up to prevent the unsafe unions is T88574372.

Reviewed By: mpage

Differential Revision: D27684041

fbshipit-source-id: e93589e60f