only inline funcs at opt level 2
Summary:
Inlining a function makes it impossible to monkeypatch it. We
want to inline `passes_sampling_rate` but we have existing tests that
expect to be able to monkeypatch it out. Long term the most robust
answer here is to emit bytecode for both the inlined code and the call
and a guard (or move all inlining to the JIT), but short term since IG runs
tests at opt level 0 and prod at opt level 2, we can make it work by
just only inlining at opt level 2.

This requires threading the opt level through to the `TypeBinder`. It
also exposed a bug where we weren't passing the opt level down to child
codegens in the code generator.

Reviewed By: DinoV

Differential Revision: D27709075

fbshipit-source-id: 6f7b5f848d