inline unboxed finals for intxx() also
Summary:
We had support for inlining the unboxed constant when we see
`unbox(FINAL_CONSTANT_INT)`. This diff adds support for doing the same
thing in the case where we have e.g. `int64(FINAL_CONSTANT_INT)`. In
general I think we want to prefer the latter since it's more explicit
about the primitive type we unbox to; we may even want to get rid of
`unbox()` entirely.

Reviewed By: DinoV

Differential Revision: D27743822

fbshipit-source-id: aca7c22060