tell compiler that box() returns an int/float
Summary:
We were setting the return type of `box()` to dynamic, which
meant that if an `-> int` function ended with
`return box(some_primitive_int)` we'd emit an unnecessary `CAST`.
Same with boxing doubles.

We know that boxing a primitive int results in a Python int, and boxing
a primitive double results in a Python float. Reflect this in the
compiler to avoid casts.

Reviewed By: DinoV

Differential Revision: D27744390

fbshipit-source-id: fdd6314706