Fix bug in JIT dict watching
Summary:
If we watch a key in a globals dict, then the builtins dict becomes
unwatchable, then we delete that key from the globals dict, we'll try to watch
the same key in the builtins dict since it's no longer shadowed (and crash,
since it's unwatchable).

We should instead reflect what would happen if the builtins dict became
unwatchable while we were already watching it: unwatch the globals key and
disable the cache.

Differential Revision: D27609308

fbshipit-source-id: 5456c8db79