minor cleanup in rewriteByteMultiply
Summary:
Turns out Dino and I were simultaneously working on native
division support, and D27772273 got there first so I won't be pushing my
work on that. But here are a few minor cleanups in `rewriteByteMultiply`
that I noticed while working on that. We already have an `in_reg` var so
we can use it instead of getting the input register again. And we may as
well size these moves appropriately; at the very least it makes the
disassembly make more sense.

Reviewed By: swtaarrs

Differential Revision: D27782743

fbshipit-source-id: 2e714fa5c2