allow inlining RAND_MAX as primitive
Summary:
This diff introduces the barest beginnings of a concept of
literal types, for NumClass only, with the only literal type in
existence so far being `__static__.RAND_MAX` and the only effect of a
literal type being that it will be inlined as a primitive if passed to
`intxx()/unbox()`.

I'm not sure how much of this we'll want to keep in this form when we
get to a fuller implementation of literal types, but this works for now.

Reviewed By: DinoV

Differential Revision: D27758164

fbshipit-source-id: e7d91eac93