Compile code objects rather than functions in HIRBuilder
Summary:
Change `HIRBuilder::BuildHIR()` to take a `PyCodeObject*` and a
globals dict rather than a `PyFunctionObject*`, and fix all the ensuing
fallout. Most of the changes are plumbing the right data around, with one
bigger change: I moved the recursive compilation check added for Static Python
from `HIRBuilder` to `_PyJIT_CompileFunction()`. This gives us the freedom to
attempt recursive compilation from any part of the JIT with the depth
protection, which we might want to use for normal functions at some point.

Differential Revision: D27343059

fbshipit-source-id: 2d7f6587ab