fix optimize level in SymbolTable.compile
Summary:
We were accidentally passing the `optimize` level passed to
`SymbolTable.compile` to the `flags` arg of the code generator. This
broke inlining in IGSRV now that we only inline at opt level 2.

Fixed this, and also updated `StaticTestBase.compile` so that in the
common case (when compiling with `StaticCodeGenerator`, and not passing
the flags to skip peephole or AST optimizer (which static tests never
do)), we go through `SymbolTable.compile` instead of `make_compiler()`,
so our tests are using more of the code path that will be used in IGSRV.

Reviewed By: DinoV

Differential Revision: D27837397

fbshipit-source-id: f2a3e04d40