Support divide directly in generated code
Summary:
Adds support for doing divide directly in the generated code instead of calling a helper.  this ends up being a significant piece of overhead in `util.knobs.passes_sampling_rate` vs Cython, so this gets us close to parity with it.

The existing binary operation / immediate pass is augmented to handle when we're dividing by an immediate.  The post-alloc pass gets a new pass which re-writes the dividend into rax.  And the register allocator is updated to know that div needs rax and rdx.

Currently we always provide an immediate zero for rdx, so there's no support for providing anything else yet.

There's some additional complications on the 8-bit version as a 16-bit value is read from rax, so we need to extend the 8-bit value into it.

Reviewed By: carljm

Differential Revision: D27772273

fbshipit-source-id: 025f843f51