Import parallel-hashmap library and provide shim for using it
Summary:
This diff adds the parallel-hashmap library from https://github.com/greg7mdp/parallel-hashmap/ at revision 259323f06fd9245136f5119ec55bff19ba088254.

This is now added to the build in `Makefile.pre.in` and a shim provided in `Jit/containers.h`. The shim provides generic types which back onto the new parallel-hashmap ones by default but can be swapped back to STL. This might be useful for debugging or experimenting with other libraries in future.

Differential Revision: D27351666

fbshipit-source-id: 7c2ff0da34