JIT: input_phy_uses should default to an empty vector
Summary:
`input_phy_uses` should default to an empty vector instead of a vector with an only element `0`.

`FIELD_DEFAULT(std::vector<int>, input_phy_uses, {})` is expaned to `std::vector<int> input_phy_uses{{}}` instead of `std::vector<int> input_phy_uses{}` as intended. Therefore, a zero is added to the vector upon initialization.

This issue does not affect the result of `getInputPhyRegUse()`, because the zero is the default value.

Reviewed By: carljm

Differential Revision: D27933110

fbshipit-source-id: 7517557871