Make async_cached_property compatible w/ IG Python version
Summary:
async_cached_property has some subtle differences from the one in distillery, and this brings them more in-line.  For starters it's checking the dictionary first although that should only be a minor perf problem.  But it also sets the value directly in the dictionary while the Cinder version goes through `setattr`.  It's also missing `fget`.

I'm not sure exactly which one of these differences is causing an issue, but the current Cinder is failing on canary with D27132904.  This combination to align the two appears to fix it.

Reviewed By: vladima

Differential Revision: D27931179

fbshipit-source-id: d686e76e0f