clear all classloader caches, between tests and subTests
Summary:
Replace `cinder.clear_vtables()` with
`cinder.clear_classloader_caches()` which clears both vtables and the
classloader caches. (I couldn't really think of a case where we'd want
to do one but not the other.)

And run this not only between each test, but also every time we call
`self.subTest()`, since actually the most likely case for classloader
caches to cause a problem is when we repeatedly create a similar module
in a loop, which will always have the same module name since it's in the
same test, and likely will also create same-named classes in that
module.

Reviewed By: DinoV

Differential Revision: D27928498

fbshipit-source-id: d58b7b9b81