avoid emitting invalid sized push/pop
Summary:
`push` and `pop` are only valid for certain sizes (16 and 64
bit, for x64), so we need to avoid emitting them at arbitrary
sizes.

There are currently two ways we can get a `push` or `pop`
in the emitted code: either a mem->mem move (our autogen
pattern for `Mm` for `kMove` uses push/pop) or an explicit
`kPush` or `kPop` instruction.

For the second case, the fix is simple; always emit `kPush` and
`kPop` as 64-bit. This doesn't hurt anything and is safe. Regalloc
copies are the only place we create `kPush` and `kPop` instructions
in LIR, so that handles the second case fully.

For the first case, remove the autogen pattern for `Mm` for `kMove`,
since it is not safe at all bit widths, and removing it ensures we
don't accidentally reintroduce a dependency on it that is only
valid at some sizes. This requires two fixes to avoid hitting this
removed case:

1. For mem->mem copies in register allocation, explicitly emit the
`kPush` and `kPop` instead of emitting a mem->mem move and
relying on codegen to turn it into push/pop.

2. Special case move instructions in `getInputPhyRegUse()` so that
if the output is a memory-indirect we require physical register input.
(A similar case in `getOutputPhyRegUse()` is not needed because
that defaults to true and is true for Move.) This covers the case
of ensuring that what would previously have been `Store` instrs
(but are now just `Move`) don't end up as mem->mem moves in a
stack-spill case. The alternative to this change could be to reintroduce
an explicit `Store` instruction that always requires physical reg input.

Differential Revision: D27912979

fbshipit-source-id: f1d748c54d