add error location context to TypedSyntaxErrors
Summary:
I'm working on a diff where I need to raise an error inside `make_generic_type`, and
I realized that currently we have no way to do this with filename and line number
information. Currently we already raise some errors in the generic-types paths, and
we just raise them badly (no filename or line number context.)

Consolidate the existing duplication of `syntax_error` methods, and introduce a new
`error_context` context manager that catches `TypedSyntaxError`, checks if it is
missing location context, and fills it in if needed. Use this context manager in generic
`visit()` method of `DeclarationsVisitor` and `TypeBinder`, as well as in a few other
places (e.g. `finish_bind`, which isn't called from a node visit, and in the node iteration
of `DeclarationsVisitor.visitClassDef`) to get more specific error locations and ensure
that no `TypedSyntaxError` should be able to escape without location info.

I didn't broadly replace calls to `TypeBinder.syntax_error(..., node)` with
`raise TypedSyntaxError(...)`, even if the `node` passed in would be the same one that
the context manager in generic `visit()` would fill in. I think we need to keep
`TypeBinder.syntax_error()` around regardless, since there are a number of places
where we call a bind method with a node that is more specific than the one we are
currently visiting (i.e. we drill into a node's attributes manually, not just by visiting.)
And if we are keeping it around anyway, I think it would be difficult and not useful to
try to minimize its use. But at the very least, we should now be able to safely just
`raise TypedSyntaxError(...)` in most cases and still get a reasonable location.

Reviewed By: DinoV

Differential Revision: D27857090

fbshipit-source-id: a310ba8225