prevent unions containing primitive types
Summary:
When we are handling primitives in the JIT, we need to know exactly
which type we are handling. Unions of primitives (and especially unions
of both primitives and python objects) aren't workable.

Part of the reason these snuck through previously is that we do
too much `isinstance(x, CType)` for detecting and special casing
primitives. This means that e.g. a union containing a primitive doesn't
raise errors in various situations where a primitive type should, since
`UnionType` is not a subclass of `CType`. If we reduced the reliance on
`isinstance` on type objects and consistently went through proxyable
methods and attributes instead, then these unions would generally have
raised type errors when used (e.g. "cannot assign primitive to
dynamic") instead of silently passing type checking.

And yes, I do use `isinstance(..., CType)` in this very diff :) Right now
we don't have a good alternative, so this is something to tackle
orthogonally.

But regardless I think the right thing to do is prevent these unions
from ever occurring in the first place.

Also fleshed out `float` and `complex` a bit, just as a side effect of needing
some non-primitive types to swap into the existing union tests that
were violating the no-primitives rule.

Reviewed By: DinoV

Differential Revision: D27945520

fbshipit-source-id: f2e810a57a