add support for cbool args/returns
Summary:
Add a bunch more missing cases for cbool to support cbool args and returns.

Also prohibit cbool arithmetic.

`TYPED_BOOL` is a bit tricky; I swapped it with `TYPED_CHAR` so that it's even,
which gives us the property that `TYPED_BOOL & TYPED_INT_SIGNED` is false,
which we want, and also that we can use `(x >> 1) & 3` to get size. This gives the
same results for all `TYPED_INT*` because they all fit in 2 bits (after stripping the
signedness bit), and it masks off the higher bits from `TYPED_BOOL` so we still
get a size of `TYPED_INT_8BIT` (i.e. 0) for it.

The swap of `TYPED_BOOL` and `TYPED_CHAR` is safe for bytecode back-compat
since we don't currently have any static modules in IGSRV using `cbool` or `char`.

In order to stay consistent that a `cbool` boxes to a Python `bool` and vice versa,
we ensure that the value for a `PRIMITIVE_LOAD_CONST` of type `TYPED_BOOL`
is Python `True` or `False`, not integer `1` or `0`. This ensures that in the interpreter
loop we load `True` or `False`.

Reviewed By: DinoV

Differential Revision: D27970464

fbshipit-source-id: f42ef0a269