Fix flaky test_pydoc test
Summary:
When `_start_server()` from `pydoc.py` returns we should get an object with a valid `docserver` attribute (*). However, all this function checks for before returning is the attribute `serving` being set to `True` and this is a race in `ServerThread.run()`.

I guess something we did recently now made the race consistently lose with ASAN runs because `test_pydoc` was consistently failing in this setup.

(*) I'm pretty sure the test is correct in this assumption as `ServerThread.stop()` also assumes the `docserver` attribute is set.

Differential Revision: D27931971

fbshipit-source-id: e731b2c938