Skip eager coroutine dispatch if await is in a different basic-block
Summary:
Unfortuantely D26854339 contains a bug - the abstract interpreter/translation loop is of the form:
```
for (auto bc_it = bc_block.begin(); bc_it != bc_block.end(); ++bc_it) {
  ...
}

```

But `emitAnyCall()` may do `bc_it++` a few times. This causes the `bc_it != bc_block.end()` to fail to catch the end of the block and process bytecode instructions until we crash. Specifically this will happen if we have Python code of the form `await (x() if y else z())`. This causes the `GET_AWAITABLE`, `LOAD_CONST`, `YIELD_FROM` sequence to be shared by both the `x()` and `y()` calls and hence be in its own basic-block. So we end up advancing the `bc_it` too far at one of the `CALL_FUNCTION` instructions.

This diff simply fixes this by just not eagerly dispatching if `GET_AWAITABLE`... are in another block.

Another solution would be to always emit the eager dispatch code if the `await` is shared. However, this might be redundant if neither branch is actually eagerly dispatchable. Checking for this seems more trouble than it's worth.

vladima also points out the inefficiency of the simple approach can be mitigated with tweaks in the bytecode compiler. E.g. turning `await (x() if y else z())` into `await x() if z else await y()`.

Reviewed By: carljm

Differential Revision: D27374561

fbshipit-source-id: 31f77dd5bf