Make cached_property not be a data descriptor
Summary:
After D27132904 rolled out we're now spending a little more time in `cached_property_get` because it's a data descriptor (https://fburl.com/scuba/pyperf_experimental/tlufzpok).  Some of this is probably time that was previously being spent elsewhere, but we can probably do a little bit better with our inline caches hitting things when the value is assigned.

This makes it a non-data descriptor which matches the implementation we had in Cython.

Reviewed By: vladima

Differential Revision: D28050937

fbshipit-source-id: 155eb9d117