Optimize calls to simple builtin functions
Summary: Just some low hanging fruit in the JIT I noticed while looking at statsd.  We have some support for this with static function when we know the target function, but there's no reason we can't generally invoke built-ins functions when we have the correct number of arguments.  This adds support for `METH_NOARGS` and `METH_O` cases.

Reviewed By: carljm

Differential Revision: D28081369

fbshipit-source-id: 6a127e368b