Fix issue with int compares
Summary: We're currently using the `type_ctx` of a compare to push down the types of the integer values.  That's not really right, the type of the compare is going to sometime be in a `bool` or `cbool` context.  So now we try and use our own context if we're not already a CInt.

Reviewed By: carljm

Differential Revision: D28129518

fbshipit-source-id: b3b667b091