Fix a crash when patching a method with primitive return type
Summary: When we resolve the return type of a function with a primitive argument we get back a NULL type as we're using `_PyClassLoader_ResolveReferenceType` which returns NULL on primitive types.  Instead we should just be using the type, and then add the special logic to see if the return value is compatible with the returned type.

Reviewed By: carljm

Differential Revision: D28168554

fbshipit-source-id: 0ca8eacc2d