Create get/set descriptor for cached properties with slots
Summary:
Turns a `cached_property` used with a descriptor into a subtype of `cached_property` which is a data descriptor.  This allows the descriptor to still be assigned to.

This is in support of the efforts to make `HttpRequest` strictified and slotified.  There are some places where `cached_property` is used but their are assignments that want to overwrite it.

We specifically don't allow this to be constructed with subtypes.  For starters the subtype probably gets `__get__` wrong by default as it'll need to explicitly look up the value from the slot, but by default people are just used to implementing something that always creates the new value.  But it also just adds complexity for dealing with the ref counts on the type objects.

Reviewed By: carljm

Differential Revision: D28229207

fbshipit-source-id: d35648ae9a