Refactor `CodeGen`
Summary:
This is a small refactoring to the `CodeGen` abstraction. Once upon
a time in a land far away (that only used C), this was going to be
the JIT backend. Since we're now using C++ and this class only generates
custom slots for type objects, update it to reflect that.

Reviewed By: swtaarrs

Differential Revision: D28012165

fbshipit-source-id: 73df7dd0dd