don't silence errors in _PyClassLoader_ResolvePrimitiveType
Summary:
In general there's no reason for us to be silencing an error
from a bad type descr in `_PyClassLoader_ResolvePrimitiveType`; it'd be
better to fail loudly if we get a bad type descr.

The one exception is the call site in `parse_slot_type`, where we could
be getting a forward reference that we can't resolve. Silence the error
at that particular callsite instead.

Reviewed By: czardoz

Differential Revision: D28368836

fbshipit-source-id: 982bbabf4d