Use PyIter_Send as a uniform way to send values into generators/coroutines
Summary:
This diff mostly backports differences between Cinder and upstreamed version of this code in CPython.
Why: In our code awaits mainly happen on coroutines. This is captured in perf data and as a result code fragments that correspond to awaiting other kinds of awaitables end up in cold code. To deal with this issue we stop treating coroutines differently and instead make them support `am_send` slot (indicated by `Py_TPFLAGS_HAVE_AM_SEND` flag on the type). As a consequence coroutines/futures/asynclazyvalues and awaitable values should end up on the same codepath

Reviewed By: mpage

Differential Revision: D28270812

fbshipit-source-id: aff4c7d316