Don't enforce return types on overrides of async functions
Summary:
Currently when an async function is patched we enforce the return type of it.  But the enforced type should really be something more like `coroutine` or `Coroutine[T]` or maybe even just some `Awaitable[T]`, instead we expect to match the expected return value, which isn't what the caller is going to expect.

This makes it so we don't enforce anything - we currently have no dependencies on the return type of an awaited call.

Reviewed By: carljm

Differential Revision: D28417142

fbshipit-source-id: 88e9134b60