don't support explicit set_type_static calls
Summary:
This line is wrong (it's reflecting on `rand` instead of
`set_type_static`), but rather than fixing it, it seems to me we should
just remove it. Is there any valid use case for importing
`set_type_static` from `__static__` and then explicitly calling it in a
static module? I think the only valid use of this function should be the
one we implicitly generate after each class is built, which doesn't
require the compiler to know about it as part of `__static__` module.

Maybe we will run into some use case someday; if so we can re-add it
here.

Reviewed By: czardoz

Differential Revision: D28445787

fbshipit-source-id: 5ef24611b8