simplify resolution APIs
Summary:
Simplify classloader type resolution APIs, since we can represent primitive types
as their `PyTypeObject*`. Get rid of the overly-complex signature of
`_PyClassLoader_ResolveType` with three out params, and simplify down to just
`_PyClassLoader_ResolveType` (which always returns `PyTypeObject*` with `optional`
out param) and `_PyClassLoader_ResolvePrimitiveType` which returns a type code
and errors on non-primitives.

The simplified signature of `_PyClassLoader_ResolveType` does mean that some
call sites now need to follow it with a call to `_PyClassLoader_GetTypeCode` if they
want to handle primitive types differently.

I decided to remove `_PyClassLoader_ResolveReferenceType` since all it adds is
extra checking to throw an error if we resolve to a primitive type, and I don't think
this is needed or worth it at any of the call sites.

Reviewed By: czardoz

Differential Revision: D28369945

fbshipit-source-id: d208f3ddff