Add shadow frame mode
Summary:
This adds a new frame mode that is intended to replace normal frames and,
ideally, no-frame mode. I ran into a showstopper in getting pure no-frame mode
working using RBP-based frame walking. The ctypes module uses libffi for performing
calls to foreign functions. Unfortunately, libffi uses some handwritten assembly
to perform those calls. It sets up a frame on the C stack that doesn't use
frame pointers, blocking our ability to unwind through it. We could potentially
get around this by using DWARF metadata instead of frame pointer unwinding,
however, we'd only be in a marginally better spot. We can't guarantee that
all C code we need to unwind through emitted the necessary DWARF metadata.
Shadow frames are a safer, slightly less efficient alternative.

Shadow frames are stack allocated by both the interpreter and JIT-compiled
functions, and are linked into a call stack. The top of the stack is stored in
the current PyThreadState. JIT-compiled functions do not allocate a `PyFrameObject`
when shadow frame mode is active.

A shadow frame has the following fields:
```
typedef struct _PyShadowFrame {
  struct _PyShadowFrame *prev;

  /*
   * If `has_pyframe` is 0, this is set to the id of the CodeRuntime of the
   * corresponding JIT-compiled function. If `has_pyframe` is 1, this value is
   * unspecified.
   */
  uint32_t code_rt_id;

  /*
   * Set to 1 if there is an active Python frame associated with this (a
   * PyFrameObject reachable from tstate->frame) or 0 otherwise.
   */
  uint32_t has_pyframe;
} _PyShadowFrame;
```

When an interpreted function is executed, the interpreter loop will push an
entry onto the shadow stack with `has_pyframe` set to `1`.

When a JIT-compiled, non-generator function is executed it will push an entry
onto the shadow stack with `code_rt_id` set to to the id of the `CodeRuntime`
associated with the function and `has_pyframe` set to `0`.

The situation is slightly more complicated for generators. A generator may have
its frame materialized (e.g. if someone accesses `gi_frame`) prior to suspending.
When it resumes, it should link the materialized frame into the Python stack. When
a JIT-compiled generator resumes, it checks to see if the generator has a `PyFrameObject`.
If so, it sets up a shadow frame in the same manner as the interpreter (`_PyJIT_GenSend`
will link the `PyFrameObject`). If not, it sets up a shadow frame in the same manner
as a non-generator function and also stores a pointer to the generator.

To materialize the Python stack for a `PyThreadState`, we walk its shadow stack
backwards. When we encounter an entry with a `has_pyframe` of `0`, we materialize
a `PyFrameObject`, link it into the Python call stack, and arrange for the frame
to be unlinked when the associated JIT-compiled function returns. If the function
was not a generator function, we arrange for it to return into a trampoline that
unlinks the Python frame before jumping to the original return address. If the
function was a generator function, we attach the frame to the associated generator
object. The epilogue has been modified to check for this scenario and unlink the
frame if it is present.

CPython also exposes the frame object used by generators (and all subtypes) via
the `gi_frame` attribute on the generator. When someone accesses `gi_frame` on
a running generator we walk the shadow stack, find the entry associated with
generator in question, and materialize a `PyFrameObject` for it. When someone
access `gi_frame` on a suspended generator, we allocate a `PyFrameObject` and attach
it to the generator. `_PyJIT_GenSend` will handle linking it into the Python stack when
the generator is resumed.

Performance on microbenchmarks is indistinguishable from no-frame. I haven't had
a chance yet to see what perf looks like in Swayze. I've left no-frame for now.
If perf is "good enough," I'd like to consolidate onto shadow frame mode and
eliminate all other frame modes entirely.

Reviewed By: swtaarrs

Differential Revision: D28346650

fbshipit-source-id: 6cfd000f7f