Track deopt events and expose to Python code
Summary:
Track how often each deopt happens, and store some additional metadata
with each one. All deopts have a string description which defaults to the HIR
opcode, but can be customized with a better string and/or add type information,
for things like type guards.

Deopts should be rare so this tracking is always on. I added a function to the
`cinderjit` module to get and clear the information, which can be used by
applications as needed.

Reviewed By: mpage

Differential Revision: D28297383

fbshipit-source-id: 29a2391eee