resolve union annotations to dynamic
Summary:
Since we don't yet have support for runtime resolving or
checking arbitrary union types (beyond the Optional degenerate case),
it's also not safe for our compiler to resolve such annotations as
anything but dynamic. In a case like:

```
def f(x: int | str) -> int:
    if isinstance(x, str):
        return 1
    return x
```

If we trust the union type of `x`, we will narrow it to `int` after the
`if` and trust (not cast) the `return x`. But in fact `CHECK_ARGS` can't
test the value of `x` at runtime, so we might have anything in `x`. Thus
we have to resolve (non-optional) union annotations as `dynamic` for
now.

It is still useful to have the full union implementation in the compiler,
because there are cases where a local union type is created, e.g. if
`x: str` and `y: int` and then `z = x or y`, the local type of `z` should be
`str | int` and this is a type we can trust and safely narrow.

This hole also caused us to emit `CHECK_ARGS` for union-typed
arguments, which were just useless checks against `object`. Added
an assert to catch if we ever try to do that again.

A lot of tests were using union annotations to check various aspects
of union behavior in the compiler; in most cases I was able to preserve
those tests by just switching to a union type synthesized from an `or`
instead of annotated.

Reviewed By: czardoz

Differential Revision: D28527662

fbshipit-source-id: 0e1ba84cf1