don't infer a declared type
Summary:
This pattern should be OK:

```
x = None
if flag:
    x = "foo"
```

Even more important, this pattern (which will compile) should not fail
at runtime because the return value of `nonstatic_func()` is not `None`:

```
x = None
if flag:
    x = nonstatic_func()
```

In order to avoid this, when the first assignment to a local is not
annotated, always set its declared type to `dynamic` so we will accept
all future assignments to it.

Module globals are handled differently; these patterns (especially the
latter) should be less common there, and local types are less useful, so
we want to infer declared types.

Previously when joining types we would fall back to declared type if
neither type was a subtype of the other. But this means that after the
first example above we have only `dynamic` for `x` where we'd rather
have `str | None`. So update type joins at branch merges to create a
union. (Unions already implement simplification of subtypes etc, so this
replaces the previous check-subtype-each-direction logic.)

Also, previously in cases of indefinite assignment we would fall back to
declared type, but this doesn't really make sense; if a local is defined
in only one branch we can just as safely keep the local type from that
branch, we don't have to fall back to declared type.

Reviewed By: czardoz

Differential Revision: D28622811

fbshipit-source-id: 54375f22ea