Add missing locks for multi-threaded compilation
Summary:
A lock was missing in `builder.cpp` which I think was due to a land-race, although I am a bit puzzled why this hasn't been breaking tests. This issue very reliably shows up for me.

There was also a missing lock writing to perf-map files which vladima pointed out. It's less surprising this wasn't found because `JIT_PERFMAP` isn't even set during test runs. Even with this set, nothing seems to show up in TSAN anyway although that probably means TSAN doesn't track thread races to file-descriptors.

The latter lock is a bit of meaty one although it should add time proportional to the total number of functions written. This seems to be about 100ms with the ~75k functions the multi-threaded test currently has.

Reviewed By: swtaarrs, vladima

Differential Revision: D28375121

fbshipit-source-id: 18df51abf9