Add link from Dockerfile to wiki for rebuild instructions and to our repo
Summary:
This has a reference to an internal Wiki, but I think this is reasonable as you need to be an employee for access to update the image anyway.

I've also added a `LABEL` referencing the Cinder repo. This seems to have allowed me to give access to update the image to everyone with write access to the Cinder repo.

Reviewed By: mpage

Differential Revision: D28697020

fbshipit-source-id: 054b48d0e8