type-check assignments to globals
Summary:
We were not correctly handling assignments to names declared as
global/nonlocal, because we would just look things up directly in
`self.decl_types`, ignoring the fact that the target name was a global.
We actually had some code in place for special-casing assignments to
globals, but it only handled `Final`, not actually type checking the
assignment.

Generalize this to a new `get_target_decl_type` method, and use this
wherever we need to get the declared type of an assignment target.

This also exposed another bug where in inferred declarations of globals,
we would infer the exact type of the RHS and not allow reassignment to a
subclass, this diff fixes that also.

There was one test in `test_cinderjit.py` exploiting the lack of type checking
on assignments to globals; modified it to do the patching from outside the
module instead.

Reviewed By: czardoz

Differential Revision: D28687293

fbshipit-source-id: e2fe74b99d