support ClassVar annotations
Summary:
For class (not instance) attributes, you are supposed to
annotate them with `typing.ClassVar[]`, like this:

```
from typing import ClassVar

class C:
    x: ClassVar[int] = 3
```

Prior to this diff, we would still try to create a slot for `x`, which
then fails at runtime because the slot conflicts with the class-level
assignment.

After this diff, we will mark the `Slot` for `x` as `is_classvar`, which
causes us to omit it from `__slots__`, emit `LOAD_ATTR` for it instead
of `LOAD_FIELD`, and know that `C.x` is of type `int`, not
`MemberDescriptorType`. Since Static Python types are frozen, it is
safe to rely on this type, even without a metaclass-level
`typed_descriptor` equivalent.

We don't get any perf benefits on loading these attributes, unlike
instance attributes. But at least after this diff we know their types
and can omit some CASTs.

Code-wise I don't love reusing `Slot` for classvars, which are not
slots, and behave pretty differently so we have to check `is_classvar`
various places. But it's also pretty convenient since we get the
construction of them for free, as well as conflict checking with other
members, etc.

Reviewed By: czardoz

Differential Revision: D28731471

fbshipit-source-id: be678eccd0